refactor: `transformStatsOverviewData` to utils
